Can just call `focus` instead of a  wtr plugin
At some point the focus/enter key based unit test was failing in webkit.
I wrote the `focusPlugin` that used the playwright `focus` method which
I thought fixed my issue. I wanted to revisit the issue and in doing so
realized that it was in fact working as expected without the plugin.
Unsure what I was doing differently at first now.